Fix tags search and hint while adding tags to test cases in TP
How to reproduce:
- Open a Test Plan
- Go to Cases tab
- Select few test cases
- click Tag -> Add Tag button
- start typing tag name
- if there are tags which match the text they will be shown
  as a drop-down hint. This was previously raising an exception.

Also introduced a slight performance optimization by using
QuerySet.values() method to fetch only the field name we're interested
in, not everything.

Note: this commit makes use of the TestCaseTag class instead of the
TestTag class in order to convert ORM queries to child -> parent
relationship, instead of parent -> child one.

Updated JavaScript files to match the backend changes and model
field names.

Note:
Removal of tags bug is #174 which is fixed in PR #177

Signed-off-by: Mr. Senko <atodorov@mrsenko.com>