Add test for #181, fails only on MySQL
Note: on MySQL the test will fail with foreign key constraint
violation b/c TestPlanEmailSettings, which has 1-1 mapping with
TestPlan does not exist!

The failure happens when we delete the TestPlan as a result of
cascading delete of a Product. There are signal handlers for TP
which send email notifications on deletion. TP.emailing
does not exist but b/c this is a property we try to create a
TestPlanEmailSettings object which will point to the deleted TP!

Signed-off-by: Mr. Senko <atodorov@mrsenko.com>