Update MOTD_LOGIN setting
- remove mention of kerberos b/c users who don't have it get confused
- remove MOTH_AUTH because it is not used at all

Updated:

While reviewing this patch, I'm not quite sure how it is useful for
daily use for users. And showing text in either h2 or h3 seems make no
sense. On the other hand, as what the comment of MOTD_LOGIN says, it
could be good to make it configurable in database so that the messages
can updated easily and no need to restart server to apply the
changes.

So, at this moment, I just update this patch to keep MOTD_LOGIN empty
and leave the decision of what message should be displayed to users who
might find it is useful for his/her use cases.

Signed-off-by: Mr. Senko <atodorov@mrsenko.com>
Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>