Restructure list of dependencies
The original structure to list dependencies in two requirements files
for runtime and development has limitation for the documentation build
in readthedocs.io. The problem is, in previous commits, django is
imported in docs/conf.py in order to generate XMLRPC API documentation
from docstrings, which requires to install django in advance. That could
be fixed easily by specifying a requirements file in the readthedocs
configuration Web page. However, a new problem is caused then.
readthedocs does not have krb5 devel package installed, which results in
build failure. As in the future, other modules would be imported in
order to publish more APIs in documentation, I have to figure out a way
to make it possible to install dependencies in readthedocs without
installing the required Kerberos package.

There are two directions for possible fixes. One is to find out a
readthedocs way to install some missing packages in order to compile
dependent packages. Another one is to just avoid installing
python-kerberos package. I choose the latter, that is to restructure the
list of dependecies eventually.

While considering the latter solution, I found Nitrate would benefit
from it to decouple the hard dependency to python-kerberos for
KerberosBackend authentication backend. So, using the setuptools
extras_require mechanism to provides packages in addition to those
required at runtime. The major changes are

* requirements/base.txt is removed. The packages are provided in
  install_requires directly in setup.py.
* kerberos is added to extras_require as optional feature that could be
  ignored by users who is not going to use Kerberos authentication.
* requirements/devel.txt is removed. The packages are split into three
  parts. One is for running tests, which is added to tests part of
  extras_require. One is for building documentation, which is added to
  docs part. One is for assisting development, which is added to
  devtools.
* Add a new requirements.txt in project root directory, which is
  generated by `pip freeze`. This file could be used to provision a
  virtualenv in order to have a consistent development environment in
  all development machines.
* Documents which mentions removed files are updated.
* Vagrantfile.example is updated to provision virtualenv from newly
  added requirements.txt.
* tox.ini is updated to use the new changes.
* test settings is updated by importing from tcms.settings.common
  instead, which can avoid to install useless packages for running
  tests.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>