Refactor object change log structure and the display
Original log is stored in a field named action. The content of action
contains both original content and the new one. This patch splits it
into two separate field in database, which are named original_value and
new_value.

Data migration is provided in this patch as well. It tries test to
migrate original value to the new fields as much as possible. However,
the actual original action content would be complicated in practice, so
the field action is not removed automatically in migration. Instead, it
is left to Nitrate site administrators to determine wether the migrated
data is correct and when is the right time to remove action eventually.

Fixes #11

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>