Reconsider setting permissions to default groups
Reason to reconsider:

While fixing tests for running against with MySQL/MariaDB, some test
swtches to subclass TransactionTestCase, then error occurs. That is,
originally, setting permissions to default groups work inside a
post_migrate signal handler, but during the teardown process of Django's
test framework for TransactionTestCase, all database tables are
truncated and trigger post_migrate again. This makes it difficult to
determine when normal migrate is executing and when that is triggered by
the test framework.

On the other hand, it seems make no sense to set permissions immediately
after migrate. The one, who is deploying Nitrate, should be able to
determine what permissions a user having perticular role has.

Solution:

So, as mentioned in Configuration document, if default permissions are
required to run Nitrate, just run the script. Then, there is still
opportunity to adjust permissions according to circumstances.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>