Convert django_comments.object_pk to integer
This is another try to convert django_comments.object_pk to a integer
column. Last time to do this in a migration, which was converted
already. And this time, data type is converted in a post_migrate signal
handler. So, it ensures the operation happens just after all tables are
created in database.

This change also benefits the PostgreSQL support, that does not work on
operations against two columns with different data types, for example in
a JOIN clause on the column.

Please note that, running tests will cause post_migrate signal is sent
every time after a test runs. There might be performance issues
probably, but it is not a big issue at this moment. If it is in the
future, consider a way to just run the signal handler just once.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>