Fix Dockerfile for released image
A custom product.py is added to separately and overwrite original
product.py before installation. This is convenient for building and
running Nitrate 4.3. The custom product.py has minimum configuration for
authentication and other stuff.

A README is also added to describe the image, whose content could be
copied to image hosting service to display for users.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>