Fix incorrect comparison with int and None for sort
While generating report "By Case-Run Tester", code sorts subtotal result by
case runs' tester user ID to ensure the order in rendered web page. However, a
test case run could have no default tester, which results in NULL is stored in
database and None is set as corresponding Python object. Hence, the sort fails
on such kind of list [1, 2, 3, None].

To fix this issue, instead of sorting in Python, the subtotal data is sorted in
database level by adding ORDER BY clause to the underlying SQL SELECT
statement.

This error happens when run Nitrate locally with Python 3.6 and PostgreSQL.

Fixes #394

In addition to issue #394, this patch also fixes same issue happening to
generate "By Plan's Tag Per Tag View" report. The solution is same. Just sort
in database by adding ORDER BY clause and avoid sorting data in Python.

Both of these two fixes are related to one issue, that is to ensure the order
of key/value pairs inside a Python dict object. The order matters, that is
because we would like to ensure to display report data in order properly to
represent the data calculated. In the beginning, OrderedDict is considered but
it is determined to just use dict finally due to since Python 3.6, dict
preserves the insertion order. There are lots of discussion about this new
feature for dict in the Internet. Since it is planned for Nitrate not to run
with Python 2, so only Python 3 is the main factor to make the decision.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>