Allow building image from version branch
docker/released/Dockerfile is refactored to allow building image from version
branch. As a result, once an image is required to be built for a specific
version, just checkout to that version by tag or branch, then `make image'.

Note that, the Dockerfile does not download any tarball from Github anymore.

With this change, it is also possible to build latest image from development
branch. By default, checkout to development branch (currently it is the develop
branch), run `make image', the default tag is 'latest'.

To build an image for a specific version, please remember to set environment
variable IMAGE_VERSION, for example:

    make image IMAGE_VERSION=4.3

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>