JS: start to test js code with qunit
qunit is used by Django to test its JavaScript code. After researching
on qunit, I found out it is so easy to understand and start to write
test code with it and the code can also be organized in a reasonable
structure easily. I'm pretty happy with qunit now. However, followings
are some points to be researched:

* mock jQuery.ajax.
* better run tests from command line in order to integrate it with CI
  process.
* show and inspect JavaScript code coverage.

Better than nothing at this moment. Just open the js-tests.html in a Web
browser and tests will run.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>