Remove cases select all
Cases select all was added to avoid loading a plan's all cases at once
in order to have a good performance when there are lots of cases for
example more than a thousand. But, this feature makes things more
complicated and it is not easy-to-use for users. This commit removes
this feature so that it would be flexible and easier for developer to
add or maintain features.

Fixes #558

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>