JS: abstract jQuery.ajax calls
There are two kinds of jQuery.ajax calls. One is to make a JSON request
which expects the server side responses data in JSON format. And another
one is to just make a normal HTTP request and server responses a piece
of HTML. These calls are abstracted into two dedicated functions which
contains some default values and behavior, which simpifies the original
calls to jQuery.ajax.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>