JS: reload case run detail pane only once
The original problem is a regression bug that is when a case run related
data is changed, for example adding a comment, the code will find all
elements which have class .expandable and trigger the click to load
those details. In addition, the original behavior before this regression
the case run detail endpoint was requested twice in order to reload the
detail.

This patch refactors the relative codebase a lot to reload the
detail only once. The relative behaviors are:

* change case run status
* add a comment to a case run
* remove a comment to a case run
* add an issue to a case run
* remove an issue from a case run
* add a link to a case run

This patch also fixes a bug that is to update the number of issues to
the case run row.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>