Make it easier to customize the settings (#815)
The settings module is able to be customized by mounting a specific
volume and providing a dedicated Python module. Any configuration within
the setting module can be customized, for example, changing the email
settings and providing site-specific useful links in the footer.

The volume to be mounted in container is ``/nitrate-config``.

The custom config module is ``nitrate_custom_conf.py``.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>