Fix django_comments.object_pk migration
Since django-contrib-comments version 2.1.0, index is added to
django_comments.object_pk. That causes the existing migration has to be
updated, otherwise it will fail due to the alter statement fails without
removing the added index.

The update is firstly to find the existing added index, remove them, and
recreate the index using btree, which is the default for PostgreSQL,
MySQL and MariaDB with InnoDB engine.

For the use inside Nitrate, integer type works well for the
django_comments.object_pk.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>