Adjust auth plugin for MySQL to run tests in CI
Since 2021/03/20 the tests failed with MySQL in GitHub CI workflow,
which reported:

  'Plugin caching_sha2_password could not be loaded:
  /usr/lib/x86_64-linux-gnu/mariadb19/plugin/caching_sha2_password.so:
  cannot open shared object file: No such file or directory

A workaround searched from the Internet which are used by many people is
to set the authentication plugin to the mysql_native_password.

However, the interesting thing is this only happens in the Github
workflow. It works well locally without requiring this change.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>