Use manage_tags view to remove tags from selected cases (#575)
In a test plan page, user is able to remove tags from selected tags.
This removal was implemented in a separate view function in testcases
app. This patch merges it into the core/ajax/manage_tags and convert the
original view function to a simple view that only responses the tag
candidates for removal.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>