bandit: Remove PYTHON_EGG_CACHE from wsgi (#1016)
bandit detects a hardcoded temporary directory path is used, but the
environment variable PYTHON_EGG_CACHE is useless for now.

The default httpd conf provided along with this codebase is configured
to run httpd process in daemon mode. For mod_wsgi, the way of configured
python-eggs cache directory is different for daemon mode and embedded
mode. So, embedded mode is required, set WSGIPythonEggs explicitly,
which can be referred in the mod_wsgi document
configuration-directives/WSGIPythonEggs.html.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>