bandit: Remove tcms.core.ajax.form (#1061)
This view function calls an insecurity exec call reported by bandit.

After inspecting the codebase, this view is only requested to update
test cases' automated state. As a final solution, this patch removes
this view function and uses an alternative way to render to dialog
content. A another benefit is there is not any HTTP request when
clicking the Automated button in the test plan page.

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>